Add files via upload
American Sign Language (ASL) is the primary language used by many deaf individuals in North America, and it is also used by hard-of-hearing and hearing individuals. The language is as rich as spoken languages and employs signs made with the hand, along with facial gestures and bodily postures.

In this project, we will train a convolutional neural network to classify images of ASL letters. After loading, examining, and preprocessing the data, we will train the network and test its performance.